pyproject.toml
[project]
name = "langchainofficialdocs"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.0.8",
    "langchain-openai>=1.0.3",
    "langchain-core>=1.0.0",
    "langgraph>=1.0.0",
    "python-dotenv>=1.2.1",
    "langchain-google-genai>=3.1.0",
    "langchain-anthropic>=1.1.0",
]

[tool.pyright]
typeCheckingMode = "basic"
# Desabilita checagens mais rígidas para trabalhar confortavelmente com frameworks dinâmicos
reportAttributeAccessIssue = "none"
reportArgumentType = "none"
reportUnknownMemberType = "none"
reportUnknownArgumentType = "none"
reportMissingTypeStubs = "none"
